move usecases folder into models
